+++ shared/simulator/src/debug.rs
@@ -36,6 +36,6 @@ pub fn emit_ship(sim: &mut Simulation, handle: ShipHandle) {
 }
 
 pub fn convert_color(c: u32) -> Vector4<f32> {
-    let extract_color = |k: i64| -> f32 { ((((c as u32) >> (k * 8)) & 0xff) as f32) / 255.0 };
+    let extract_color = |k: i64| -> f32 { (((c >> (k * 8)) & 0xff) as f32) / 255.0 };
     vector![extract_color(2), extract_color(1), extract_color(0), 1.0]
 }
